Reduced manifest for initial v2 app plus incomplete script for its maint
The v2 IsraTv app needs lots of work.  On previous app I was able to
make the man menu land on the primary category which was intentionally
in the middle of the manifes so we would have "interesting" categories
to the right and left.  With the grid used for v2 the category we land
on first is the first one listed.  So doing some reordering and
reducing of optins till can get a better organization of all the
offerings.

Additionally v2 does not yet deal with background work needed to take
a MAKO or dailymotion URL to produce the actual HLS (m3u8) URL, so
filtering those out.

Initial script needs some work that for now I will just manually edit
in subsequent commit:
    - Need to remove whole menue entries that are left without any
      streams.
    - Will probably add some new metadata with system message and/or
      version check.